+++ lib/src/applied_mathematics/statistic/exceptions/mean_exception.dart
@@ -1,0 +1,11 @@
+/// An exception that trows when performing unaccepted opetation or operation under unaccepted values
+class MeanException implements Exception {
+  /// Creates instance of [MeanException]
+  MeanException(this.message);
+
+  /// Reason of exception
+  final String message;
+
+  @override
+  String toString() => '$message';
+}